+++ src/backend/mod.rs
@@ -1,8 +1,11 @@
 pub mod svg;
 pub mod terminal;
 
-#[cfg(feature = "png")]
+#[cfg(feature = "raster")]
 pub mod png;
+
+#[cfg(feature = "raster")]
+pub mod raster;
 
 #[cfg(feature = "pdf")]
 pub mod pdf;